syslog: ensure first batch of UDP packets fits in ARP queue
Since the network stack needs to do an initial ARP query in order
to retrieve the MAC address to be used for packets with a given
destination, if many such packets are sent to the same destination
before a response to the ARP query is received, the ARP driver
queues only a limited number (currently 3) of them. As a result, if
the user application prints many messages in short succession
during startup and expects these messages to be delivered to a
syslog server, from the initial burst of messages, only the 3 most
recent ones are delivered.
To work around the above issue, this commit changes the syslog klib
so that the number of UDP packets sent in the first batch does not
exceed the size of the ARP queue.
In syslog_udp_send(), the `seconds` local variable is being renamed
to `secs` in order to fix a clash with the seconds() function being
called when expanding the SYSLOG_FLUSH_INTERVAL definition.

Partially solves #1836.