unixsock_close: close rather than deallocate sockets in conn_q
The existing code in unixsock_close() could create a condition whereby a
reference for a deleted child socket would remain in the process file
descriptor vector. This could materialize as a use-after-free and crash upon
process exit.

The remedy presented here allows the child socket to be created upon
connection, but defers the allocation of an fd until the child is accepted
from the listening socket. When a listening socket is closed, its child
sockets are closed via the usual method rather than deleted outright.