add vector value and tag type
This introduces a new vector type to the tagged value system. Previously,
vectors were represented as tuples with vector indices encoded as keys. With
these changes, a tagged vector, allocated via allocate_tagged_vector(), may be
accessed and manipulated using the standard vector functions found in
src/runtime/vector.h. As such, vector elements may be consumed, popped,
deleted, etc., whereas the previous treatment of vectors as tuples would make
such operations cumbersome, requiring an element to be deleted and re-inserted
if its index would change after an operation.

Note that these changes require a change to the tuple encoding format: an
additional bit has been added to the encoding type, allowing for a total of
four encoding types.

Also note that such changes would require a corresponding change to the
filesystem writing code in ops. It may be worth incrementing TFS_VERSION in
tfs_internal.h and allowing old encodings to be decoded (and types converted
to the new arrangement) when older versions are found.