riscv: fix read-only remapping of memory addresses
The existing definition of PAGE_PROT_FLAGS for riscv prevents the
kernel from being able to remove the write permission on an
existing mapping.
This commit fixes the above issue and adds a test case to the
in-kernel page table tests to verify that remapping from read-write
mode to read-only mode works.
In addition, the .bss.ro_after_init linker section is now aligned
to a page boundary, otherwise when this section is remapped as
read-only the remapping touches other sections that should not be
touched.