On-demand mapping: invoke pending fault completion on failure
If the kernel cannot map a page being requested after a page fault,
the pending fault structure that is allocated to handle the fault
should have its completion invoked, otherwise if another fault
occurs on the same page the newly faulting context is added to the
list of dependents of the pending fault and then is suspended
indefinitely.
In addition, any error status tuple should be properly deallocated
by the pending fault completion.