runtime: add touch_memory() function
This function touches all pages covered by a given address range,
and can be used to ensure that a given set of pages is resident in
memory (or to check whether an address range can be mapped). The
fault_in_user_memory() function has been refactored to call
touch_memory().
The definition of PAGELOG, PAGESIZE and PAGEMASK has been moved to
architecure-specific machine.h files.