Kernel contexts: add process context type
This context type is used to run code that is triggered by the user
process via a syscall but is executed asynchronously (and thus
cannot be associated to the syscall context that triggered it).
Syscall contexts and process contexts are now sub-types of a "Unix
context", which is an abstract context type that can block on a
block queue; all blockq-related fields in the thread struct have
been moved to the unix_context struct, and the blockq code has been
changed to operate on unix contexts instead of threads; this allows
I/O operations running in a process context to block just like
those running in a syscall.
All context types that execute code in kernel mode are now
sub-types of an abstract kernel_context type; this allows
de-duplicating common code shared by different context types, e.g.
in functions such as init_kernel_context().
The Unix fault handler is no longer associated to a thread, but to
the user process as a whole; when a fault occurs in kernel mode
(caused by invalid access to user memory), the kernel does not send
a signal to userspace and does not abandon any ongoing syscall
(which may not only leak resources, but also leave other things in
an inconsistent state, e.g. by failing to decrement refcounts or
release locks); instead, the next commit will implement error
handling via the addition of an error frame to the kernel_context
type.
All operations triggered by a syscall that have an asynchronous
completion instantiate this completion as a contextual closure, so
that it can run in the appropriate context (which is typically a
syscall context or a process context). To easily instantiate
closures that should be associated to a given context, the
closure_from_context() macro has been defined. File descriptor
closures that take a completion argument now have a context
(instead of thread) parameter, so that their implementation for
the various file descriptor types can create an appropriate
contextual closure to handle asynchronous completions.
The thread parameter has been removed from the io_completion
closure type, because such completions may not be associated to
any user thread, and their associated (Unix) context is always the
current context.
In unix/aio.c and unix/io_uring.c, process contexts are used to
run asynchronous completions, thus the code added by commits
d55147c89446b5acbc1fc5cd960e1b045b649cfc and
04c96c1d8aa7fc3edbcd18813e308b30f5d2ef0b is no longer needed and
has been removed.
All on-demand page requests following a page fault triggered in
kernel mode are no longer associated to a user thread, but to the
user process; the thread_demand_file_page closure embedded in the
thread struct has been replaced by the demand_file_page() function
that is be called directly by demand_filebacked_page().
The signalfd code has been amended so that signals retrieved via a
read() are associated to the thread that issues the read request
(instead of the thread that created the signalfd); this allows
proper handling of read() invocations from a non-syscall context,
where there is no current thread for which pending signals should
be retrieved, and is consistent with the Linux implementation (for
Linux, an SFD_TASK flag has been proposed in
https://lwn.net/Articles/805862/ to tie a signalfd to the thread
that created it, but this proposal has not made it to mainline).