Disallow overwriting of program and interpreter files by default
With this change, the user application by default is not allowed to
overwrite the program (and interpreter, if present) binary file.
This default behavior can be overridden by inserting a
"program_overwrite" attribute in the root tuple of the manifest.
In the write runtime test, the "exec_protection" flag has been
removed in order to test that the executable file by default cannot
be overwritten even without exec protection.
File access permissions are implemented in a generic way so that it
is possible to mark an arbitrary file as read-only by inserting a
"readonly" attribute in the file tuple in the manifest.