Unix: implement RELATIME (relaxed update of file access time)
With this policy, when a file is accessed, the file access time is
updated only if its current value is less than or equal to the file
modification time (or if it is more than 1 day old). This is the
default behavior on Linux, and makes file reads more efficient by
avoiding frequent file metadata updates.