vmxnet3: fix allocation size for rx queue completion descriptors
The memory allocated for the rx queue completion ring was half the
size required for the number of completion descriptors as set in
the comp_ring_len field of the vmxnet3_rxq_shared struct. This was
causing memory corruption as the vmxnet3 device would write past
the memory range allocated for the completion ring.

Closes https://github.com/nanovms/ops/issues/1454