vmxnet3: disable Large Receive Offload (LRO) feature
If the LRO feature is enabled, when the vmwnet3 device coalesces
multiple input TCP packets into a single packet, it does not update
the checksum in the TCP header of the large packet (because it can
indicate that all the checksums are correct by setting the
appropriate fields in the completion descriptor); therefore, when
the packet is processed by the lwIP TCP stack, the checksum
calculation fails, and the packet is discarded. This causes a high
performance penalty, since the application needs to wait for the
discarded packet to be retransmitted by the remote peer. On the
other hand, checksum calculation by lwIP should be kept enabled
because lwIP does not support disabling the calculation on a
per-packet basis, and it is not possible for the driver to know
beforehand whether the checkusm for incoming packets will be
calculated by the device.
To fix the above issue, this change disables the LRO feature; in
addition, the RX checksum feature is also disabled because we
cannot exploit it, since we cannot avoid the checkusm calculation
in lwIP.
The MTU size is being updated to match the size of the RX buffers
(the vmxnet3 device discards any incoming packets larger than the
configured MTU value).