TFS log: do not reference log refcount if TLOG_READ_ONLY is defined
Having a refcount in the TFS log structure is only needed to
implement the log compaction procedure; therefore, this refcount is
only initialized if the TLOG_READ_ONLY pre-processor define is not
present. However, the current code accesses the refcount
unconditionally in the functions that open and close a log
extension, which can cause the invocation of a non-initialized
completion thunk. This has been observed when running the UEFI
bootloader on ESXi.
This change fixes the above issue by putting the code that accesses
the refcount under `#ifndef TLOG_READ_ONLY`.