x86_64: install_gdt64_and_tss: set up segment registers
When a GDT is loaded, the CPU registers that contain a segment
selector need to be set up so that they refer to an appropriate
segment in the GDT; failure to do so may result in a general
protection fault being raised when executing instructions such as
`iretq`.
When the kernel boots via the PC platform bootloader, these
registers are set up by the bootloader according to a GDT whose
kernel code and data segment indices equal the corresponding
indices in the kernel GDT; but when the kernel boots via UEFI, the
GDT being used by the UEFI boot code may have segments at different
indices.
This change modifies the install_gdt64_and_tss() assembly function
so that CPU registers are correctly set up according to the GDT
defined in the kernel; this decouples the kernel GDT from the
bootloader GDT, and properly handles a potential mismatch between
the kernel GDT and the UEFI GDT.