x86_64: add support for retrieving ACPI tables from UEFI firmware
During ACPI initializaton, the current code in the kernel looks for
the RSDP (which contains pointers to the ACPI tables) in the BIOS
ROM and EDBA memory areas. In UEFI machines, the UEFI firmware that
runs during boot may place the RSDP at an arbitrary position in
memory, outside the address ranges used by the kernel, which causes
the kernel to fail to initialize the ACPI subsystem.

This change modifies the UEFI bootloader so that it retrieves the
RSDP location from the UEFI firmware and puts it in a region; then,
the kernel retrieves the RSDP when scanning the regions.
This fixes ACPI initializtion in ESXi instances booted via UEFI.