vmxnet3: enable I/O and memory access in the PCI device
In ESXi instances booted via UEFI, access to memory and I/O space
is disabled by default in the vmxnet3 PCI device, which causes the
driver to fail to initialize the device.
This change amends the vmxnet3 driver to enable memory and I/O
access via the command register in the PCI device configuration.