X86_64: PVH entry point: add stack memory configuration
The pvh_start32 assembly code is a 32-bit entry point for the
kernel, used by hypervisors such as QEMU (with the microvm machine
type) and cloud-hypervisor
(https://github.com/cloud-hypervisor/cloud-hypervisor).
The current code was missing the initialization of the stack, which
prevents the kernel from booting under cloud-hypervisor.
This change adds the code to set the stack pointer register and to
map the memory area to be used as initial stack during boot; this
allows booting under cloud-hypervisor.

Closes #1838.