TFS: extend(): fix calculation of range to be allocated
In the extend() function, the `blocks` range is not relative to the
starting block where the extent is allocated in the storage medium;
therefore, the expression `ex->start_block + blocks.end` is wrong,
and leads to a larger-than-needed allocation request, which wastes
storage space and can lead to a suboptimal file extent map, with
more extents than necessary.
This change fixes the above issue.