TFS log: fix struct log refcounting
In the current code, a log extension bumps the refcount of the
relevant struct log until the extension is closed. This causes the
log refcount to never reach zero, because there always is at least
one extension (i.e. the current extension) open, with the end
result that after a log compaction the resources associated to the
log being retired are leaked.
This change fixes the above issue by moving the increment and
decrement of the log refcount from the code that opens and closes
an extension to the code that handles log flushing to disk (which
is the only asynchronous operation that can potentially complete
after a log is compacted).