TFS log: prevent multiple parallel log compaction operations
With the current code, it can happen that a log is deemed eligible
for compaction even before it is actually linked to the
filesysytem, i.e. before retiring the log that is being compacted.
This causes issues if a newly started compaction completes before
the completion of the previous compaction, because when the
previous compaction finally completes the cleanup code can access
metadata tuples that have been deallocated by the new compaction.
This change fixes the above issue by replacing the `compacting`
boolean flag in the log structure with an enum that indicates the
current state of a log (INIT, LINKED, COMPACTING, FAILED); a log
is now eligible for compaction only when in the LINKED state. The
`failed` flag is also being removed since a failure now results in
the log switching to the FAILED state.