TFS storage heap: replace id heap with rangemap-based heap
Memory allocation required to operate the storage heap for a given
TFS volume increases linearly with the size of the volume, with 1
bit of memory required for every 512-byte sector in the volume; for
example, a 1-TB volume requires 256 MB of memory for its storage
heap (which can become up to double that amount due to how buffer
extension works, where the capacity of a buffer is doubled when its
current capacity is not enough).
To decrease memory consumption and prevent large memory allocations
when new disk space is requested (which can be a problem when the
kernel is under memory pressure), this change replaces the id heap
with a rangemap-based heap, which represents each allocated disk
area with a rangemap node. The number of nodes in the rangemap is
minimized by merging adjacent ranges into a single node.