Page cache: deallocate page structs when draining the cache
Keeping a struct pagecache_page in memory after the page has been
evicted means that the total memory occupied by the pagecache
increases with the numbeor of pages that have been allocated since
boot, even if those pages have been evicted. This can be a problem
in applications that read or write large files (and/or a high
number of files) in the filesystem.
This change modifies the pagecache_page_release_locked() function
so that it can deallocate the struct pagecache_page in addition to
deallocating the page memory. Struct dellocation is done on a best-
effort basis, because a struct being deleted must be removed from
set of pages of the corresponding pagecache node, which needs to be
locked while keeping the pagecache lock. The
pagecache_page_release_locked() function takes a boolean argument
indicating whether to attempt deallocation of the struct; the
argument is set to false in cases where the pagecache_page struct
pointer is accessed after the function call (e.g. when traversing
the list of pages of a node), or when a pagecache node is already
locked.
In order to be able to reclaim the memory allocated for
pagecache_page structs, a new caching_heap (objcache) is now used
instead of the general heap for allocating these structs. In
addition, pagecache_drain() tries to delete (again, on a
best-effort basis) any free pages that couldn't be deleted when
they were deallocated.