mm_service(): use a dynamic threshold for memory cleaning
Since a certain level of memory fragmentation is unavoidable when
memory is dynamically allocated and deallocated on the various
heaps, a given amount of free memory does not imply the ability to
allocate a chunk of contiguous memory, and the likelihood of an
amount of free memory to allow allocating a chunk of memory of a
certain size decreases with increasing total memory size. For this
reason, having a fixed threshold to decide when triggering memory
cleaning does not scale well to instances with large memory size.
This change modifies the mm_service() function so that memory
cleaning is triggered when the amount of free memory goes below a
threshold that depends on the total memory size; this makes memory
cleaning more effective on instances with large memory size, and
solves issues that have been observed where the kernel is unable to
allocate, say, 2 MB even after memory cleaning has been executed
successfully.