Closures: allow re-using closure structs with different functions
Until now, a closure function was part of the definition of the
closure, meaning that it was not possible to attach different
functions to different instantiations of a closure.
With this change, the above limitation is removed (at least for
closures embedded in other structs); a closure struct can be
declared in a header file and instantiated in different places with
different functions: the init_closure_func() macro initializes a
closure with an arbitrary function (which must still have the
appropriate signature), and the closure_func_basic() macro is used
to define the function to be associated to the closure (which must
be a basic closure, i.e. its struct must not have any
closure-specific fields).
The unix_internal.h header file now defines closure structs for the
different closures associated to a file descriptor, and the inotify
and io_uring implementations use these definitions instead of
defining their own closure structs.