eventfd: use generic file descriptor closures and fix memleak
With this change, the eventfd implementation embeds generic file
descriptor closure structs within struct efd, instead of allocating
closures from the heap. This optimizes memory usage, minimizes the
number of memory dereferences in the closure functions, and
improves cache locality; in addition, it fixes a memory leak due to
a missing deallocation of the edge_trigger_handler closure when an
eventfd file descriptor is closed.