Unix domain sockets: fix memory leaks
With this change, the Unix domain socket implementation embeds
generic file descriptor closure structs within struct unixsock,
instead of allocating closures from the heap. This optimizes memory
usage, minimizes the number of memory dereferences in the closure
functions, and improves cache locality; in addition, it fixes
memory leaks due to a missing deallocation of the sg_read, sg_write
and ioctl closures when a Unix socket file descriptor is closed.