Network sockets: use generic file descriptor closures
With this change, the network socket implementation embeds generic
file descriptor closure structs within struct netsock, instead of
allocating closures from the heap. This optimizes memory usage,
minimizes the number of memory dereferences in the closure
functions, and improves cache locality; in addition, it fixes a
memory leak due to a missing deallocation of the sg_write closure
when a socket file descriptor is closed.