Real-time clock: add support for running without an RTC peripheral
Nanos can run on platforms where the RTC peripheral is read-only
(e.g. riscv-virt), is not initialized with the system time from the
host (e.g. cloud-hypervisor), or is not available at all (e.g.
AWS Firecracker). This change set allows the system time to be
retrieved during boot from an alternative source (e.g. the wall
clock time made available by the paravirtualized clock), and allows
changing the system time even when the RTC peripheral is
unavailable (in this case, the time update will not persist across
a VM reset).
The register_platform_clock_now() function now takes an additional
parameter with the value of the offset between the system time and
the monotonic time: if such offset is available (i.e. non-zero), it
is used to initialize the rtc_offset field of the VDSO structure,
otherwise the kernel falls back to trying to derive the offset from
the RTC time. On platforms with the pvclock peripheral, the RTC
offset is available via the wall clock time.
The x86 RTC driver is being amended to be able to handle cases
where the RTC is not available: in these cases, rtc_gettimeofday()
returns 0, and rtc_settimeofday() is a no-op.

Closes #1869