Klibs: gcp: refactor logging code to allow reuse for metrics
This refactoring (no functional changes intended) renames the
`gcp_log_setup*` functions to `gcp_setup*` and changes them to be
able to handle the case where logging is disabled; in addition, the
`gcp_log_connect` function is being replaced by a more generic
`gcp_connect` function; the code that allocates a closure for
asynchronous execution in the DNS callback function is being
removed because it is no longer necessary since when the global
lwIP lock has been removed.
These changes are made in preparation for the next commit that will
add support for sending metrics for the GCP monitoring service.