Klibs: gcp: add support for sending metrics to monitoring service
This change enhances the gcp klib so that it can now be configured
for sending memory usage metrics to the GCP monitoring service,
thus emulating the GCP ops agent.
The ID of the running instance and the zone where it is running
(which are necessary to be able to send API requests to the
monitoring server) are retrieved from the instance metadata server.
The metrics being sent are related to memory usage (see
https://cloud.google.com/monitoring/api/metrics_opsagent#agent-memory),
and more specifically are the `bytes_used` and `bytes_percent`
metric types; for each type, a value is sent for each of the
"cached", "free" and "used" states.
Sending of memory metrics is enabled by inserting a "metrics" tuple
in the "gcp" configuration tuple; it is possible to modify the time
interval at which metrics are sent by inserting an "interval"
attribute in the "metrics" tuple, with the interval value expressed
in seconds; the default (and minimum allowed) interval value is 60
seconds.