Signal user process to exit on poweroff request
When the hypervisor signals nanos to shut down, nanos should first ask the
user process to terminate so that it has a chance to quiesce state. This
change introduces a unix_shutdown function that sends SIGTERM to the
user process and sets a compile-time configurable timer to force a
shutdown. If the user process handles the SIGTERM then the vm will
shutdown as soon as the process exits. If the user process does not
exit by the expiration of the shutdown timeout, the kernel will proceed
with shutdown. The unix_shutdown() function should be called from
places that process shutdown notifications from the hypervisor, like
the acpi power button event. The default signal handler has been
modified not to dump signal and thread information when receiving a
SIGTERM.