virtio_net: use minimal backing page sizes for caches
Given that there is limited queueing within the network stack, we can
reasonably predict that only up to virtqueue_entries(vn->rxq) rx buffers will
need to be cached - and similarly for tx handler completions. This adjusts the
objcache parent pagesizes to accommodate only these requirements, regardless
of whether the machine is in a 'low memory' configuration.