virtio PCI: dynamically assign MSI-X entries
A virtio PCI device that supports MSI-X interrupts does not
necessarily have (<N> + 1) MSI-X entries (where <N> is the number
of virtqueues it implements): it may have an MSI-X table with a
smaller number of entries. For example, the "vhost-user-vsock-pci"
device implemented in QEMU version 5.2.0 has 3 virtqueues (rx, tx,
and events) and 3 MSI-X entries.
The existing code, which assumes that there is an available MSI-X
slot for each virtqueue plus 1 slot for configuration change
interrupts, may fail to initialize a virtqueue if the above
assumption does not hold.
This change replaces the fixed assignment of MSI-X slots with a
dynamic allocation, so that the kernel can initialize all of the
<N> virtqueues supported by a device, provided that the MSI-X table
has at least <N> entries and the configration change interrupt is
not used. This fixes the "cannot configure virtqueue MSI-X vector"
error that is triggered when trying to set up MSI-X for the event
virtqueue of the vhost-user-vsock-pci device in QEMU 5.2.0.