Unix: add vsock support
This change set adds support for the vsock socket family, which
allows a program running in a virtual machine to communicate with
programs running in the host or in other virtual machines.
The vsock transport mechanism currently supported is the mechanism
implemented by virtio socket devices, used in both QEMU and AWS
Firecracker. The current vsock implementation supports stream-type
sockets.

Closes #1876.