Add additional tests for optional nanos features
Several nanos features and their code is not used unless explicitly enabled
in the manifest, and the normal runtime tests only cover some of these. This
change introduces a python script to exercise some of these other optional
features to ensure they continue to function when nanos changes. An additional
variable parameter is added to the makefile to allow for the easy addition of
manifest options through the make command line. This is used to modify the
behavior of existing runtime tests as a one-off to test the optional features.

This initial set of tests covers the manifest options that are both not
currently covered and also relatively easy to verify. These include the
netconsole feature, missing_files, syscall_summary, coredumplimit,
notrace and tracelist, debugsyscalls, ltrace, and noaslr. The script has been
added to run in the build-pc ci script as well.