x86_64 compiler options: disable use of MMX instructions
Since the kernel code does not contain any floating point
operations, it does not save and restore the FPU state when
switching between user mode and kernel mode. However, in the x86
architecture, MMX registers are aliases for the floating
point registers, thus execution of any MMX instruction might modify
the FPU state.
The gcc 12 compiler generates MMX instructions when compiling the
AcpiUtInitGlobals() function in the acpica code; this is causing
the FPU to enter a state that results in an invalid operation
FPU exception and leads to wrong computation results when the user
program executes floating point instructions.
This change adds the `-mno-mmx` flag to the compiler options for
x86, so that generation of MMX instructions is avoided.

Closes #1884.