ACPI: fix GED interrupt handler registration on x86_64
This changes moves the AcpiOsInstallInterruptHandler() and
AcpiOsRemoveInterruptHandler() functions from arch-specific code to
common code, and adds a new arch-specific function called
acpi_register_irq_handler(). This eliminates code duplicated across
architectures, and fixes registration of the interrupt handler for
the generic event device (GED), which now works not only on aarch64
but also on x86_64.