ACPI: decouple power button probing from _S5 object probing
In the existing code, the ACPI power button is probed only if an
_S5 object cannot be found in the ACPI tables. However, the power
button and the _S5 object can exist independently from each other,
and thus should be probed separately.
This change moves probing of the power button from
acpi_powerdown_init() to init_acpi(), so that the power button is
probed regardless of the presence of an _S5 object; this allows the
kernel to support the power button on platforms (such as
cloud-hypervisor) where the _S5 object is present.