ACPI powerdown: add support for Sleep Control Register
The Sleep Control Register can be used alternatively to the PM1
Control Register to power down a machine.
This change modifies the _S5 object probing code so that it detects
which of the 2 alternative powerdown mechanisms is supported in a
given platform, and installs the appropriate halt_handler closure.
This allows the kernel to power down a VM running on
cloud-hypervisor.