prevent infinite recursion when encoding tuples
This change modifies the tuple encoding functions to only encode a tuple's
content once per call, thus avoiding infinite recursion in cases where there
exists a cycle in a tuple structure (for instance, if a ".." directory entry
is mistakenly encoded and the current directory metadata is revisited as a
child of the parent directory). It achieves this by passing a table of visited
tuples through the call stack, recording each encoded tuple (immediate or
reference) in the table. Future attempts to encode any tuple found in the
table (within the span of the encode_tuple(), encode_eav() or encode_value()
call) will result in a reference to the tuple being encoded with a zero
attribute count. The tuple_test unit test has been updated accordingly to
reflect this behavior, and a check has been added to test the encoding of a
tuple that contains a reference to itself.